+++ packages/contracts/deploy/new/10_framework/50_dao-factory.ts
@@ -4,9 +4,9 @@ import {DeployFunction} from 'hardhat-deploy/types';
 import {getContractAddress} from '../../helpers';
 
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
-  const {deployments, getNamedAccounts} = hre;
+  const {deployments, ethers} = hre;
   const {deploy} = deployments;
-  const {deployer} = await getNamedAccounts();
+  const [deployer] = await ethers.getSigners();
 
   // Get `DAORegistry` address.
   const daoRegistryAddress = await getContractAddress('DAORegistry', hre);
@@ -18,7 +18,7 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   );
 
   await deploy('DAOFactory', {
-    from: deployer,
+    from: deployer.address,
     args: [daoRegistryAddress, pluginSetupProcessorAddress],
     log: true,
   });